Finish namespacing orchestrate commands
